backend/app/ai/requirements.txt
opencv-python==4.8.1.78
mediapipe==0.10.8
numpy==1.26.2
python-socketio[client]==5.10.0
